Modularized frontend and Flask backend ready for deployment
