+++ facial-frontend/access.html
@@ -1,0 +1,108 @@
+<!DOCTYPE html>
+<html lang="en">
+
+<head>
+  <meta charset="UTF-8" />
+  <title>User Access - Cybersecurity Facial</title>
+  <style>
+    body {
+      background: linear-gradient(135deg, #0f2027 0%, #2c5364 100%);
+      min-height: 100vh;
+      margin: 0;
+      display: flex;
+      align-items: center;
+      justify-content: center;
+      font-family: 'Segoe UI', Arial, sans-serif;
+    }
+
+    .access-container {
+      background: rgba(20, 30, 48, 0.95);
+      border-radius: 18px;
+      box-shadow: 0 8px 32px 0 rgba(31, 38, 135, 0.37);
+      padding: 40px 32px 32px 32px;
+      width: 350px;
+      display: flex;
+      flex-direction: column;
+      align-items: center;
+      border: 1.5px solid #00ffe7;
+    }
+
+    .access-container h2 {
+      color: #00ffe7;
+      margin-bottom: 18px;
+      letter-spacing: 1px;
+      text-shadow: 0 0 8px #00ffe7aa;
+    }
+
+    .access-container input {
+      width: 100%;
+      padding: 12px;
+      margin: 10px 0;
+      border: none;
+      border-radius: 8px;
+      background: #232b3a;
+      color: #00ffe7;
+      font-size: 1em;
+      outline: none;
+      box-shadow: 0 0 0 1.5px #00ffe733;
+      transition: box-shadow 0.2s;
+    }
+
+    .access-container input:focus {
+      box-shadow: 0 0 0 2.5px #00ffe7;
+    }
+
+    .access-container button {
+      width: 100%;
+      padding: 14px;
+      margin-top: 16px;
+      border: none;
+      border-radius: 8px;
+      background: linear-gradient(90deg, #00ffe7 0%, #007cf0 100%);
+      color: #181f2a;
+      font-size: 1.1em;
+      font-weight: bold;
+      cursor: pointer;
+      box-shadow: 0 2px 8px #00ffe744;
+      letter-spacing: 1px;
+      transition: background 0.2s, color 0.2s;
+    }
+
+    .access-container button:hover {
+      background: linear-gradient(90deg, #007cf0 0%, #00ffe7 100%);
+      color: #fff;
+    }
+
+    #video {
+      margin-top: 18px;
+      border-radius: 10px;
+      box-shadow: 0 4px 16px #00ffe755;
+      border: 2px solid #00ffe7;
+      display: none;
+    }
+
+    #msg {
+      color: #00ffe7;
+      margin-top: 10px;
+      font-size: 1em;
+      text-shadow: 0 0 6px #00ffe7aa;
+    }
+  </style>
+</head>
+
+<body>
+  <div class="access-container">
+    <h2>Secure Access</h2>
+    <form id="accessForm" autocomplete="off">
+      <input name="username" placeholder="Username" required>
+      <input name="ssnn" placeholder="SSNN" required>
+      <button type="button" onclick="validateAndStartCamera()">Activate camera</button>
+    </form>
+    <video id="video" width="320" height="240" autoplay></video>
+    <canvas id="canvas" width="320" height="240" style="display:none;"></canvas>
+    <p id="msg"></p>
+  </div>
+  <script src="assets/access.js"></script>
+</body>
+
+</html>
