+++ facial-frontend/assets/index.js
@@ -1,0 +1,8 @@
+document.addEventListener('DOMContentLoaded', function () {
+  document.getElementById('register-btn').onclick = function () {
+    window.location.href = 'register.html';
+  };
+  document.getElementById('access-btn').onclick = function () {
+    window.location.href = 'access.html';
+  };
+});
