+++ facial-frontend/documents.html
@@ -1,0 +1,94 @@
+<!DOCTYPE html>
+<html lang="en">
+
+<head>
+  <meta charset="UTF-8" />
+  <title>Your Documents - Cybersecurity Facial</title>
+  <style>
+    body {
+      margin: 0;
+      height: 100vh;
+      background: linear-gradient(135deg, #0f2027 0%, #2c5364 100%);
+      display: flex;
+      flex-direction: column;
+      align-items: center;
+      justify-content: center;
+      font-family: 'Segoe UI', Arial, sans-serif;
+      color: #00ffe7;
+      text-align: center;
+    }
+
+    .main-container {
+      background: rgba(20, 30, 48, 0.95);
+      border-radius: 18px;
+      box-shadow: 0 8px 32px 0 rgba(31, 38, 135, 0.37);
+      padding: 40px 32px 32px 32px;
+      width: 370px;
+      display: flex;
+      flex-direction: column;
+      align-items: center;
+      border: 1.5px solid #00ffe7;
+    }
+
+    h1 {
+      font-size: 2.2em;
+      margin-bottom: 18px;
+      letter-spacing: 1px;
+      text-shadow: 0 0 8px #00ffe7aa;
+      color: #00ffe7;
+    }
+
+    .doc-list {
+      color: #fff;
+      margin-top: 20px;
+      font-size: 1.1em;
+      text-align: left;
+    }
+
+    .footer {
+      margin-top: 18px;
+      color: #00ffe7aa;
+      font-size: 0.95em;
+      letter-spacing: 1px;
+    }
+
+    .btn-logout {
+      width: 100%;
+      padding: 14px 0;
+      margin-top: 24px;
+      font-size: 1.1em;
+      border: none;
+      border-radius: 8px;
+      background: linear-gradient(90deg, #00ffe7 0%, #007cf0 100%);
+      color: #181f2a;
+      font-weight: bold;
+      cursor: pointer;
+      box-shadow: 0 2px 8px #00ffe744;
+      letter-spacing: 1px;
+      transition: background 0.2s, color 0.2s;
+    }
+
+    .btn-logout:hover {
+      background: linear-gradient(90deg, #007cf0 0%, #00ffe7 100%);
+      color: #fff;
+    }
+  </style>
+</head>
+
+<body>
+  <div class="main-container">
+    <h1>Welcome! Your Documents</h1>
+    <div class="doc-list">
+      <ul>
+        <li>Document 1.pdf</li>
+        <li>Document 2.pdf</li>
+        <li>Document 3.pdf</li>
+      </ul>
+    </div>
+    <button class="btn-logout" onclick="window.location.href='/'">Logout</button>
+    <div class="footer">© 2025 Cybersecurity Facial System</div>
+  </div>
+  <script src="assets/documents.js"></script>
+</body>
+
+</html>
