+++ facial-frontend/index.html
@@ -1,0 +1,92 @@
+<!DOCTYPE html>
+<html lang="en">
+
+<head>
+  <meta charset="UTF-8" />
+  <title>Facial Recognition - Cybersecurity</title>
+  <style>
+    body {
+      margin: 0;
+      height: 100vh;
+      background: linear-gradient(135deg, #0f2027 0%, #2c5364 100%);
+      display: flex;
+      flex-direction: column;
+      align-items: center;
+      justify-content: center;
+      font-family: 'Segoe UI', Arial, sans-serif;
+      color: #00ffe7;
+      text-align: center;
+    }
+
+    .main-container {
+      background: rgba(20, 30, 48, 0.95);
+      border-radius: 18px;
+      box-shadow: 0 8px 32px 0 rgba(31, 38, 135, 0.37);
+      padding: 40px 32px 32px 32px;
+      width: 370px;
+      display: flex;
+      flex-direction: column;
+      align-items: center;
+      border: 1.5px solid #00ffe7;
+    }
+
+    h1 {
+      font-size: 2.5em;
+      margin-bottom: 18px;
+      letter-spacing: 1px;
+      text-shadow: 0 0 8px #00ffe7aa;
+      color: #00ffe7;
+    }
+
+    img {
+      width: 120px;
+      margin-bottom: 18px;
+      border-radius: 12px;
+      box-shadow: 0 4px 16px #00ffe755;
+      border: 2px solid #00ffe7;
+    }
+
+    .btn {
+      width: 100%;
+      padding: 15px 0;
+      margin: 10px 0;
+      font-size: 1.2em;
+      border: none;
+      border-radius: 8px;
+      cursor: pointer;
+      transition: background 0.3s, color 0.3s;
+      background: linear-gradient(90deg, #00ffe7 0%, #007cf0 100%);
+      color: #181f2a;
+      font-weight: bold;
+      box-shadow: 0 2px 8px #00ffe744;
+      letter-spacing: 1px;
+    }
+
+    .btn:hover {
+      background: linear-gradient(90deg, #007cf0 0%, #00ffe7 100%);
+      color: #fff;
+    }
+
+    .footer {
+      margin-top: 18px;
+      color: #00ffe7aa;
+      font-size: 0.95em;
+      letter-spacing: 1px;
+    }
+  </style>
+</head>
+
+<body>
+  <div class="main-container">
+    <img
+      src="https://images.unsplash.com/photo-1506368265446-e4dcbd427a91?ixlib=rb-4.0.3&auto=format&fit=crop&w=600&q=80"
+      alt="Logo" />
+    <h1>Cybersecurity Facial Recognition</h1>
+    <button class="btn" id="register-btn">Register</button>
+    <button class="btn" id="access-btn">Access</button>
+    <div class="footer">© 2025 Cybersecurity Facial System</div>
+  </div>
+  <script src="assets/index.js"></script>
+</body>
+
+</html>
